Checkpoint: Drawer component
There are two things we are tracking, the scroll progress of the user when they scroll before the drawer is maximized, and then the scroll progress of the inner content.

any scroll action should contribute to either of these regardless if they hover the drawer or parent. But, we must handle these cases correctly.

If the user scrolls before the drawer is maximized, all scroll events should contribute to making the drawer fill the page (scrollY).
Once the drawer is fully open, we can set isDrawerOpen boolean to true.
Then, if this is the case, any new scroll events should scroll the content of the inner drawer {children}.
The only way the user can close the drawer is if they are at the top of the content.
We can use a boolean and some scrollYProgress of the inner children content to control this.
It should be at the top.
If the user is at the top of the content and we detect scroll up, then it should begin closing the drawer (scrollY)