Add context for visitor
Context handles if the user has visited the site before, and this is being used to conditionally control the animation of the navigation.

If the user has visited the site before, the navigation will not animate.
If the user has not visited the site before, the navigation will animate.
The content on the home page will have different animation timings based on the user's visit history.
ie. 1.5s delay on the home page content if the navigation is animating.