Refactor widget grid for more flexibility.
Previously, we were rendering the widget grid by importing specific widgets into the global widget grid.
This is limiting because if we want to use the widget grid on another page, we will have to import every possible widget that grid will render into that single component thats being reused in multiple pages.

I've refactored the widget grid types to use native sanity types so we arent regernating types manually, as well as creating a registry that is defined on the page the widget is rendered on.
We pass in a registry of _type to component, and then pass that into the widget grid. The widget grid then renders the component that is defined in the registry.

This allows us to use the widget grid on any page, and only import the widgets that are needed for that specific page.